Bind controls to data in Blazor, support changing size and removing pizzas

+++ interacting-with-data-blazor/BlazingPizza/Models/Pizza.cs
@@ -23,7 +23,7 @@ namespace BlazingPizza
 
         public decimal GetBasePrice()
         {
-            return Size / DefaultSize * Special!.BasePrice;
+            return ((decimal)Size / (decimal)DefaultSize) * Special!.BasePrice;
         }
 
         public decimal GetTotalPrice()
